cb - finished testing for navbar
